+++ corehq/apps/reports/filters/dates.py
@@ -30,13 +30,3 @@ class DatespanFilter(BaseReportFilter):
             'datespan': self.datespan,
             'timezone': self.timezone.zone,
         }
-
-
-class DaterangeFilter(DatespanFilter):
-    @property
-    def filter_context(self):
-        context = super(DaterangeFilter, self).filter_context
-        context.update({
-            'use_daterange': True,
-        })
-        return context
